added types to api.ts
