+++ src/api.ts
@@ -1,26 +1,39 @@
 const API = "https://api.hackillinois.org";
+import { MethodType, FileType, RegistrationType } from "./types";
 
-function request(method, endpoint, body) {
+function request(method: MethodType, endpoint: string, body?: any) {
+    const getToken = (): string => {
+        const token = isAuthenticated();
+        if (token === null) {
+            throw new Error("token is null");
+        }
+        return token;
+    };
     return fetch(API + endpoint, {
         method,
         headers: {
-            Authorization: sessionStorage.getItem("token"),
+            Authorization: getToken(),
             "Content-Type": "application/json",
         },
         body: JSON.stringify(body),
-    }).then((res) => {
-        if (res.ok) {
-            return res.json();
-        }
-        throw Error(res);
-    });
+    })
+        .then((res: Response) => {
+            if (res.status === 200) {
+                return res.json();
+            } else {
+                throw new Error("response status code not 200");
+            }
+        })
+        .catch((err: Error) => {
+            console.error(err);
+        });
 }
 
 export function isAuthenticated() {
     return sessionStorage.getItem("token");
 }
 
-export function authenticate(to) {
+export function authenticate(to: string) {
     if (process.env.REACT_APP_TOKEN) {
         sessionStorage.setItem("token", process.env.REACT_APP_TOKEN);
     } else {
@@ -53,11 +66,15 @@ export function getRolesSync() {
     return [];
 }
 
-export function getRegistration(role) {
+export function getRegistration(role: string) {
     return request("GET", `/registration/${role}/`);
 }
 
-export function register(isEditing, role, registration) {
+export function register(
+    isEditing: boolean,
+    role: string,
+    registration: RegistrationType
+) {
     const method = isEditing ? "PUT" : "POST";
     return request(method, `/registration/${role}/`, registration);
 }
@@ -66,12 +83,12 @@ export function getRSVP() {
     return request("GET", "/rsvp/");
 }
 
-export function rsvp(isEditing, registration) {
+export function rsvp(isEditing: boolean, registration: RegistrationType) {
     const method = isEditing ? "PUT" : "POST";
     return request(method, "/rsvp/", registration);
 }
 
-export function uploadFile(file, type) {
+export function uploadFile(file: File, type: FileType) {
     return request("GET", `/upload/${type}/upload/`)
         .then((res) =>
             fetch(res[type], {
@@ -84,7 +101,10 @@ export function uploadFile(file, type) {
             if (res.ok) {
                 return res;
             }
-            throw Error(res);
+            throw new Error("response did not have status 200");
+        })
+        .catch((err: Error) => {
+            console.error(err);
         });
 }
 
@@ -108,11 +128,14 @@ export function getMentorTimeslots() {
     );
 }
 
-export function setMentorTimeslots(data) {
+export function setMentorTimeslots(data: Record<string, string>) {
     return request("PUT", "/upload/blobstore/", {
         id: "mentor-timeslots",
         data,
-    }).then((res) => res.data);
+    })
+        .then((res: Response) => res.json())
+        .then((res: any) => res.data)
+        .catch((err: Error) => console.log(err));
 }
 
 export function getEvents() {
